spoc_ppt/requirements.txt
Pillow==11.1.0
playwright==1.48.0
